Touch up app bar
